requirements.txt
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cssselect==1.2.0
cycler==0.12.1
Django==5.0.4
esprima==4.0.1
fonttools==4.51.0
idna==3.7
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
llvmlite==0.42.0
lxml==5.2.1
numba==0.59.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
plotly==5.21.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tenacity==8.2.3
tzdata==2024.1
urllib3==2.2.1
validator-collection==1.5.0
webencodings==0.5.1
